PlayNote function moved to own file
