--- Urho3D/splash_build/CMakeFiles/Progress/4
@@ -1,1 +1,0 @@
-empty
