--- Urho3D/splash_build/CMakeFiles/Progress/count.txt
@@ -1,1 +1,0 @@
-5
